+++ src/main/java/com/alwo/model/ShipmentMethod.java
@@ -6,6 +6,7 @@ import lombok.Setter;
 import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
+import java.math.BigDecimal;
 
 @Entity
 @Setter
@@ -20,9 +21,9 @@ public class ShipmentMethod {
     @NotEmpty
     @Size(min = 2, max = 10)
     private String shipmentMethod;
-    private double shipmentCost;
+    private BigDecimal shipmentCost;
 
-    public ShipmentMethod(@NotEmpty @Size(min = 2, max = 10) String shipmentMethod, double shipmentCost) {
+    public ShipmentMethod(@NotEmpty @Size(min = 2, max = 10) String shipmentMethod, BigDecimal shipmentCost) {
         this.shipmentMethod = shipmentMethod;
         this.shipmentCost = shipmentCost;
     }
